implement some command and options

+++ index.mjs
@@ -1,15 +1,86 @@
 import { existsSync, readdirSync } from 'node:fs';
 import { mkdir, writeFile, readFile } from 'node:fs/promises';
 import { deserialize, serialize } from 'node:v8';
+import { parseArgs } from 'node:util';
+import { createInterface } from 'node:readline/promises';
 
 import { cert, initializeApp } from 'firebase-admin/app';
 import { getRemoteConfig } from 'firebase-admin/remote-config';
 
-const credPath =
-  process.env['FIREBASE_CREDENTIALS'];
+const help = `example:
+  node index.mjs checkout --json FIREBASE_CREDENTIALS_JSON_PATH
+  node index.mjs validate --json FIREBASE_CREDENTIALS_JSON_PATH
+  node index.mjs publish --json FIREBASE_CREDENTIALS_JSON_PATH
+  node index.mjs publish --dryrun --json FIREBASE_CREDENTIALS_JSON_PATH
+`;
 
+function parse() {
+
+  const options = {
+    json: {
+      type: 'string',
+    },
+    debug: {
+      type: 'boolean',
+      default: false,
+    },
+    force: {
+      type: 'boolean',
+      short: 'f',
+      default: false,
+    },
+    dryrun: {
+      type: 'boolean',
+      default: false,
+    },
+  };
+
+  try {
+    return parseArgs({ options, allowPositionals: true });
+  } catch(error) {
+    console.error(error.message);
+    console.error();
+    console.error(help);
+    process.exit(1);
+  }
+}
+
+const { values, positionals } = parse()
+
+const command = {
+  checkout: false,
+  validate: false,
+  publish: false,
+};
+if (positionals.length != 1) {
+  console.error(`Error: Required one command`);
+  console.error();
+  console.error(help);
+  process.exit(1);
+} else {
+  switch(positionals[0]) {
+    case "checkout":
+      command.checkout = true;
+      break;
+    case "validate":
+      command.validate = true;
+      break;
+    case "publish":
+      command.validate = true;
+      command.publish = true;
+      break;
+    default:
+      console.error(`Error: Unknown command '${positionals[0]}'`);
+      console.error();
+      console.error(help);
+      process.exit(1);
+      break;
+  }
+}
+
+const credPath = values.json
 if (!credPath) {
-  console.error('Error: FIREBASE_CREDENTIALS env is not set');
+  console.error('Error: --json option is not set');
   process.exit(1);
 }
 
@@ -17,8 +88,8 @@ initializeApp({ credential: cert(credPath) });
 
 const remoteConfig = getRemoteConfig();
 const template = await remoteConfig.getTemplate();
-if (process.env["DEBUG"]) {
-  console.log("[current template json]");
+if (values.debug) {
+  console.log("Debug: current template json");
   console.log(JSON.stringify(template, null, 2));
 }
 
@@ -26,31 +97,33 @@ if (process.env["DEBUG"]) {
 const makeFiles = async (dir, parameters) => {
     for (const [key, parameter] of Object.entries(parameters)) {
       const file = `${dir}/${key}`;
-      console.log(key, parameter, ` => ${file}`);
-      try {
-        await writeFile(file, JSON.stringify(parameter, null, 2));
-      } catch (error) {
-        console.error(error)
-        process.exit(1);
+      if (values.debug) {
+        console.log(`Debug: checkout ${file}`);
+      }
+      if (!values.dryrun) {
+        try {
+          await writeFile(file, JSON.stringify(parameter, null, 2));
+        } catch (error) {
+          console.error(error)
+          process.exit(1);
+        }
       }
     }
-
 };
 
 const parametersDir = './parameters';
-if (!existsSync(parametersDir)) {
-    try {
-      await mkdir(parametersDir);
-    } catch (error) {
-      console.error(error);
-      process.exit(1);
-    }
-
-  makeFiles(parametersDir, template.parameters);
-}
-
 const parameterGroupsDir = './parameterGroups';
-if (!existsSync(parameterGroupsDir)) {
+if (command.checkout) {
+  if (!existsSync(parametersDir)) {
+      try {
+        await mkdir(parametersDir);
+      } catch (error) {
+        console.error(error);
+        process.exit(1);
+      }
+  }
+  makeFiles(parametersDir, template.parameters);
+
   for (const [key, parameterGroup] of Object.entries(template.parameterGroups)) {
     const groupDir = `${parameterGroupsDir}/${key}`;
     if (!existsSync(groupDir)) {
@@ -101,23 +174,38 @@ for (const group of groups) {
   }
 }
 
-try {
-  remoteConfig.validateTemplate(newTemplate)
-} catch (error) {
-  console.error(error);
-  process.exit(1);
-}
-
-// Update remote config
-if (process.env['DEBUG']) {
-  console.log("");
-  console.log("[created template json]");
-  console.log(JSON.stringify(newTemplate, null, 2));
-} else {
+if (command.validate) {
   try {
-    const result = await remoteConfig.publishTemplate(newTemplate)
+    remoteConfig.validateTemplate(newTemplate)
   } catch (error) {
     console.error(error);
     process.exit(1);
+  }
+}
+
+// Update remote config
+if (command.publish && values.debug) {
+  console.log("[Debug] created template json");
+  console.log(JSON.stringify(newTemplate, null, 2));
+}
+if (command.publish && !values.dryrun) {
+  let s = "y";
+  if (!values.force) {
+    const ri = createInterface({
+      input: process.stdin,
+      output: process.stdout
+    });
+    s = await ri.question("publish? [y]: ");
+    ri.close();
+  }
+  if (s == "y") {
+    console.log("publishing...");
+    try {
+      const result = await remoteConfig.publishTemplate(newTemplate);
+      console.log(`complete! (version ${result.version.versionNumber})`);
+    } catch (error) {
+      console.error(error);
+      process.exit(1);
+    }
   }
 }
